requirements.txt
Django==4.0.3
djangorestframework==3.13.1
psycopg2==2.9.3
django-environ==0.8.1
celery==5.2.3
gevent==21.12.0
uvicorn==0.17.6
channels==3.0.4